Matrix-defunct {Matrix}	R Documentation

Defunct Functions in Package Matrix

Description

The functions or variables listed here are no longer part of Matrix as they are no longer needed.

Usage

## Defunct in 1.3-3
rBind(..., deparse.level = 1)
cBind(..., deparse.level = 1)

Details

These either are stubs reporting that they are defunct, or have been removed completely (apart from being documented here).

rBind and cBind were provided for R versions older than 3.2.0 as substitutes for the base analogues rbind and cbind, which at that time were not “S4-aware” and so could not be used to vertically or horizontally concatenate "Matrix" objects together with traditional matrices and vectors.

See Also

Defunct, base-defunct, Matrix-deprecated

[Package Matrix version 1.6-2 Index]