dgeMatrix-class {Matrix}	R Documentation

Class "dgeMatrix" of Dense Numeric (S4 Class) Matrices

Description

A general numeric dense matrix in the S4 Matrix representation. dgeMatrix is the “standard” class for dense numeric matrices in the Matrix package.

Objects from the Class

Objects can be created by calls of the form new("dgeMatrix", ...) or, more commonly, by coercion from the Matrix class (see Matrix) or by Matrix(..).

Slots

x:

Object of class "numeric" - the numeric values contained in the matrix, in column-major order.

Dim:

Object of class "integer" - the dimensions of the matrix - must be an integer vector with exactly two non-negative values.

Dimnames:

a list of length two - inherited from class Matrix.

factors:

Object of class "list" - a list of factorizations of the matrix.

Methods

The are group methods (see, e.g., Arith)

Arith

signature(e1 = "dgeMatrix", e2 = "dgeMatrix"): ...

Arith

signature(e1 = "dgeMatrix", e2 = "numeric"): ...

Arith

signature(e1 = "numeric", e2 = "dgeMatrix"): ...

Math

signature(x = "dgeMatrix"): ...

Math2

signature(x = "dgeMatrix", digits = "numeric"): ...

matrix products %*%, crossprod() and tcrossprod(), several solve methods, and other matrix methods available:

Schur

signature(x = "dgeMatrix", vectors = "logical"): ...

Schur

signature(x = "dgeMatrix", vectors = "missing"): ...

chol

signature(x = "dgeMatrix"): see chol.

coerce

signature(from = "dgeMatrix", to = "lgeMatrix"): ...

coerce

signature(from = "dgeMatrix", to = "matrix"): ...

coerce

signature(from = "matrix", to = "dgeMatrix"): ...

colMeans

signature(x = "dgeMatrix"): columnwise means (averages)

colSums

signature(x = "dgeMatrix"): columnwise sums

diag

signature(x = "dgeMatrix"): ...

dim

signature(x = "dgeMatrix"): ...

dimnames

signature(x = "dgeMatrix"): ...

eigen

signature(x = "dgeMatrix", only.values= "logical"): ...

eigen

signature(x = "dgeMatrix", only.values= "missing"): ...

norm

signature(x = "dgeMatrix", type = "character"): ...

norm

signature(x = "dgeMatrix", type = "missing"): ...

rcond

signature(x = "dgeMatrix", norm = "character") or norm = "missing": the reciprocal condition number, rcond().

rowMeans

signature(x = "dgeMatrix"): rowwise means (averages)

rowSums

signature(x = "dgeMatrix"): rowwise sums

t

signature(x = "dgeMatrix"): matrix transpose

See Also

Classes Matrix, dtrMatrix, and dsyMatrix.

[Package Matrix version 1.3-3 Index]