unpackedMatrix-class {Matrix}	R Documentation

Virtual Class "unpackedMatrix" of Unpacked Dense Matrices

Description

Class "unpackedMatrix" is the virtual class of dense matrices in "unpacked" format, storing all m*n elements of an m-by-n matrix. It is used to define common methods for efficient subsetting, transposing, etc. of its proper subclasses: currently "[dln]geMatrix" (unpacked general), "[dln]syMatrix" (unpacked symmetric), "[dln]trMatrix" (unpacked triangular), and subclasses of these, such as "dpoMatrix", "Cholesky", and "BunchKaufman".

Slots

Dim, Dimnames:

as all Matrix objects.

Extends

Class "denseMatrix", directly. Class "Matrix", by class "denseMatrix", distance 2. Class "mMatrix", by class "Matrix", distance 3. Class "replValueSp", by class "Matrix", distance 3.

Methods

pack

signature(x = "unpackedMatrix"): ...

unpack

signature(x = "unpackedMatrix"): ...

isSymmetric

signature(object = "unpackedMatrix"): ...

isTriangular

signature(object = "unpackedMatrix"): ...

isDiagonal

signature(object = "unpackedMatrix"): ...

t

signature(x = "unpackedMatrix"): ...

diag

signature(x = "unpackedMatrix"): ...

diag<-

signature(x = "unpackedMatrix"): ...

Author(s)

Mikael Jagan

See Also

pack and unpack; its virtual "complement" "packedMatrix"; its proper subclasses "dsyMatrix", "ltrMatrix", etc.

Examples

showClass("unpackedMatrix")
showMethods(classes = "unpackedMatrix")

[Package Matrix version 1.5-3 Index]