Unused-classes {Matrix}	R Documentation

Virtual Classes Not Yet Really Implemented and Used

Description

iMatrix is the virtual class of all integer (S4) matrices. It extends the Matrix class directly.

zMatrix is the virtual class of all complex (S4) matrices. It extends the Matrix class directly.

Examples

showClass("iMatrix")
showClass("zMatrix")

[Package Matrix version 1.2-2 Index]