BunchKaufman-methods {Matrix}	R Documentation

Bunch-Kaufman Decomposition Methods

Description

The Bunch-Kaufman Decomposition of a square symmetric matrix A is A = P LDL' P' where P is a permutation matrix, L is unit-lower triangular and D is block-diagonal with blocks of dimension 1 x 1 or 2 x 2.

Usage

BunchKaufman(x, ...)

Arguments

x 	 a symmetric square matrix.

... 	 potentially further arguments passed to methods.

Methods

Currently, only methods for dense numeric symmetric matrices are implemented.

x = "dspMatrix"

...

x = "dsyMatrix"

...

These rely on the Lapack routines dsptrf and dsytrf, respectively.

See Also

The resulting class, BunchKaufman. Related decompositions are the LU, lu, and the Cholesky, chol (and for sparse matrices, Cholesky).

Examples

data(CAex)
dim(CAex)
isSymmetric(CAex)# TRUE
CAs <- as(CAex, "symmetricMatrix")
if(FALSE) # no method defined yet for *sparse* :
   bk. <- BunchKaufman(CAs)
## does apply to *dense* symmetric matrices:
bkCA <- BunchKaufman(as(CAs, "denseMatrix"))
bkCA

[Package Matrix version 0.999375-30 Index]