Hilbert {Matrix}	R Documentation

Generate a Hilbert matrix

Description

Generate the n by n symmetric Hilbert matrix. Because these matrices are ill-conditioned for moderate to large n, they are often used for testing numerical linear algebra code.

Usage

Hilbert(n)

Arguments

n
a non-negative integer.

Value

the n by n symmetric Hilbert matrix as a "dpoMatrix" object.

See Also

the class dpoMatrix

Examples

Hilbert(6)

[Package Matrix version 1.3-3 Index]