MatrixFactorization-class {Matrix}	R Documentation

Class "MatrixFactorization" of Matrix Factorizations

Description

The class "MatrixFactorization" is the virtual (super) class of (potentially) all matrix factorizations of matrices from package Matrix.

Objects from the Class

A virtual Class: No objects may be created from it.

Slots

Dim:

Object of class "integer" - the dimensions of the original matrix - must be an integer vector with exactly two non-negative values.

Methods

dim

(x) simply returns x@Dim, see above.

expand

signature(x = "MatrixFactorization"): this has not been implemented yet for all matrix factorizations. It should return a list whose components are matrices which when multiplied return the original Matrix object.

show

signature(object = "MatrixFactorization"): simple printing, see show.

solve

signature(a = "MatrixFactorization", b= .): solve A x = b for x; has not been implemented for all factorizations, but e.g., for CHMfactor, see there.

See Also

classes inheriting from "MatrixFactorization", such as CHMfactor, LU, and sparseQR.

Examples

showClass("MatrixFactorization")

[Package Matrix version 0.999375-30 Index]