MatrixFactorization-class {Matrix}	R Documentation

Class "MatrixFactorization" of Matrix Factorizations

Description

The class "MatrixFactorization" is the virtual (super) class of (potentially) all matrix factorizations of matrices from package Matrix.

The class "CholeskyFactorization" is the virtual class of all Cholesky decompositions from Matrix (and trivial sub class of "MatrixFactorization").

Objects from the Class

A virtual Class: No objects may be created from it.

Slots

Dim:

Object of class "integer" - the dimensions of the original matrix - must be an integer vector with exactly two non-negative values.

Methods

dim

(x) simply returns x@Dim, see above.

expand

signature(x = "MatrixFactorization"): this has not been implemented yet for all matrix factorizations. It should return a list whose components are matrices which when multiplied return the original Matrix object.

show

signature(object = "MatrixFactorization"): simple printing, see show.

solve

signature(a = "MatrixFactorization", b= .): solve A x = b for x; see solve-methods.

See Also

classes inheriting from "MatrixFactorization", such as LU, Cholesky, CHMfactor, and sparseQR.

Examples

showClass("MatrixFactorization")
getClass("CholeskyFactorization")

[Package Matrix version 1.3-3 Index]