MatrixFactorization-class {Matrix}	R Documentation

Virtual Class "MatrixFactorization" of Matrix Factorizations

Description

MatrixFactorization is the virtual class of factorizations of m-by-n matrices A, having the general form

P1 * A * P2 = A1 * ... * Ap

or (equivalently)

A = P1' * A1 * ... * Ap * P2'

where P1 and P2 are permutation matrices. Factorizations requiring symmetric A have the constraint P2 = P1', and factorizations without row or column pivoting have the constraints P1 = Im and P2 = In, where Im and In are the m-by-m and n-by-n identity matrices.

CholeskyFactorization, BunchKaufmanFactorization, SchurFactorization, LU, and QR are the virtual subclasses of MatrixFactorization containing all Cholesky, Bunch-Kaufman, Schur, LU, and QR factorizations, respectively.

Slots

Dim

an integer vector of length 2 giving the dimensions of the factorized matrix.

Dimnames

a list of length 2 preserving the dimnames of the factorized matrix. Each element must be NULL or a character vector of length equal to the corresponding element of Dim.

Methods

determinant

signature(x = "MatrixFactorization", logarithm = "missing"): sets logarithm = TRUE and recalls the generic function.

dim

signature(x = "MatrixFactorization"): returns x@Dim.

dimnames

signature(x = "MatrixFactorization"): returns x@Dimnames.

dimnames<-

signature(x = "MatrixFactorization", value = "NULL"): returns x with x@Dimnames set to list(NULL, NULL).

dimnames<-

signature(x = "MatrixFactorization", value = "list"): returns x with x@Dimnames set to value.

length

signature(x = "MatrixFactorization"): returns prod(x@Dim).

show

signature(object = "MatrixFactorization"): prints the internal representation of the factorization using str.

solve

signature(a = "MatrixFactorization", b = .): see solve-methods.

unname

signature(obj = "MatrixFactorization"): returns obj with obj@Dimnames set to list(NULL, NULL).

See Also

The virtual class compMatrix of factorizable matrices.

Classes extending CholeskyFactorization, namely Cholesky, pCholesky, and CHMfactor.

Classes extending BunchKaufmanFactorization, namely BunchKaufman and pBunchKaufman.

Classes extending SchurFactorization, namely Schur.

Classes extending LU, namely denseLU and sparseLU.

Classes extending QR, namely sparseQR.

Generic functions Cholesky, BunchKaufman, Schur, lu, and qr for computing factorizations.

Generic functions expand1 and expand2 for constructing matrix factors from MatrixFactorization objects.

Examples

showClass("MatrixFactorization")

[Package Matrix version 1.6-5 Index]