SparseM-conversions {Matrix}	R Documentation

Sparse Matrix Coercion from and to those from package SparseM

Description

Methods for coercion from and to sparse matrices from package SparseM are provided here, for ease of porting functionality to the Matrix package, and comparing functionality of the two packages. All these work via the usual as(., "<class>") coercion,

  as(from, Class)

Methods

from = "matrix.csr", to = "dgRMatrix"

...

from = "matrix.csc", to = "dgCMatrix"

...

from = "matrix.coo", to = "dgTMatrix"

...

from = "dgRMatrix", to = "matrix.csr"

...

from = "dgCMatrix", to = "matrix.csc"

...

from = "dgTMatrix", to = "matrix.coo"

...

from = "sparseMatrix", to = "matrix.csr"

...

from = "matrix.csr", to = "dgCMatrix"

...

from = "matrix.coo", to = "dgCMatrix"

...

from = "matrix.csr", to = "Matrix"

...

from = "matrix.csc", to = "Matrix"

...

from = "matrix.coo", to = "Matrix"

...

See Also

The documentation in CRAN package SparseM, such as SparseM.ontology, and one important class, matrix.csr.

[Package Matrix version 1.3-3 Index]