all-methods {Matrix}	R Documentation

"Matrix" Methods for Functions all() and any()

Description

The basic R functions all and any now have methods for Matrix objects and should behave as for matrix ones.

Methods

all

signature(x = "Matrix", ..., na.rm = FALSE): ...

any

signature(x = "Matrix", ..., na.rm = FALSE): ...

all

signature(x = "ldenseMatrix", ..., na.rm = FALSE): ...

all

signature(x = "lsparseMatrix", ..., na.rm = FALSE): ...

Examples

M <- Matrix(1:12 +0, 3,4)
all(M >= 1) # TRUE
any(M < 0 ) # FALSE
MN <- M; MN[2,3] <- NA; MN
all(MN >= 0) # NA
any(MN <  0) # NA
any(MN <  0, na.rm = TRUE) # -> FALSE

[Package Matrix version 0.999375-30 Index]