compMatrix-class {Matrix}	R Documentation

Class "compMatrix" of Composite (Factorizable) Matrices

Description

Virtual class of composite matrices; i.e., matrices that can be factorized, typically as a product of simpler matrices.

Objects from the Class

A virtual Class: No objects may be created from it.

Slots

factors:

Object of class "list" - a list of factorizations of the matrix. Note that this is typically empty, i.e. list(), initially and is updated automagically whenever a matrix factorization is computed.

Dim

,

Dimnames:

inherited from the Matrix class, see there.

Extends

Class "Matrix", directly.

[Package Matrix version 0.999375-30 Index]