denseMatrix-class {Matrix}	R Documentation

Virtual Class "denseMatrix" of All Dense Matrices

Description

This is the virtual class of all dense (S4) matrices. It is the direct superclass of ddenseMatrix, ldenseMatrix

Extends

class "Matrix" directly.

Slots

exactly those of its superclass "Matrix".

Methods

Use showMethods(class = "denseMatrix", where = "package:Matrix") for an overview of methods.

Extraction ("[") methods, see [-methods.

See Also

colSums, kronecker, and other such methods with own help pages.

Its superclass Matrix, and main subclasses, ddenseMatrix and sparseMatrix.

Examples

showClass("denseMatrix")

[Package Matrix version 1.3-4 Index]