dgRMatrix-class {Matrix}	R Documentation

Sparse Compressed, Row-oriented Numeric Matrices

Description

The dgRMatrix class is a class of sparse numeric matrices in the compressed, sparse, row-oriented format. In this implementation the non-zero elements in the rows are sorted into increasing column order.

Note: Currently, the column-oriented sparse classes, e.g., dgCMatrix, are preferred and better supported in the Matrix package.

Objects from the Class

Objects can be created by calls of the form new("dgRMatrix", ...).

Slots

j:

Object of class "integer" of length nnzero (number of non-zero elements). These are the column numbers for each non-zero element in the matrix.

p:

Object of class "integer" of pointers, one for each row, to the initial (zero-based) index of elements in the row.

x:

Object of class "numeric" - the non-zero elements of the matrix.

Dim:

Object of class "integer" - the dimensions of the matrix.

Methods

coerce

signature(from = "matrix", to = "dgRMatrix")

coerce

signature(from = "dgRMatrix", to = "matrix")

coerce

signature(from = "dgRMatrix", to = "dgTMatrix")

diag

signature(x = "dgRMatrix"): returns the diagonal of x

dim

signature(x = "dgRMatrix"): returns the dimensions of x

image

signature(x = "dgRMatrix"): plots an image of x using the levelplot function

See Also

the dgCMatrix class.

[Package Matrix version 0.999375-30 Index]