dtpMatrix-class {Matrix}	R Documentation

Packed triangular dense matrices

Description

The "dtpMatrix" class is the class of triangular, dense, numeric matrices in packed storage. The "dtrMatrix" class is the same except in nonpacked storage.

Objects from the Class

Objects can be created by calls of the form new("dtpMatrix", ...) or by coercion from other classes of matrices.

Slots

uplo:

Object of class "character". Must be either "U", for upper triangular, and "L", for lower triangular.

diag:

Object of class "character". Must be either "U", for unit triangular (diagonal is all ones), or "N"; see triangularMatrix.

x:

Object of class "numeric". The numeric values that constitute the matrix, stored in column-major order. For a packed square matrix of dimension d * d, length(x) is of length d(d+1)/2 (also when diag == "U"!).

Dim,Dimnames:

The dimension (a length-2 "integer") and corresponding names (or NULL), inherited from the Matrix, see there.

Extends

Class "ddenseMatrix", directly. Class "triangularMatrix", directly. Class "dMatrix" and more by class "ddenseMatrix" etc, see the examples.

Methods

%*%

signature(x = "dtpMatrix", y = "dgeMatrix"): Matrix multiplication; ditto for several other signature combinations, see showMethods("%*%", class = "dtpMatrix").

coerce

signature(from = "dtpMatrix", to = "dtrMatrix")

coerce

signature(from = "dtpMatrix", to = "matrix")

determinant

signature(x = "dtpMatrix", logarithm = "missing"): ...

determinant

signature(x = "dtpMatrix", logarithm = "logical"): ...

diag

signature(x = "dtpMatrix"): ...

norm

signature(x = "dtpMatrix", type = "character"): ...

norm

signature(x = "dtpMatrix", type = "missing"): ...

rcond

signature(x = "dtpMatrix", norm = "character"): ...

rcond

signature(x = "dtpMatrix", norm = "missing"): ...

solve

signature(a = "dtpMatrix", b = "missing"): ...

solve

signature(a = "dtpMatrix", b = "matrix"): ...

t

signature(x = "dtpMatrix"): ...

unpack

signature(x = "dtpMatrix"): ...

See Also

Class dtrMatrix

Examples

showClass("dtrMatrix")

example("dtrMatrix-class")
(p1 <- as(T2, "dtpMatrix"))
str(p1)
(pp <- as(T, "dtpMatrix"))
stopifnot(length(p1@x) == 3, length(pp@x) == 3,
          p1 @ uplo == T2 @ uplo, pp @ uplo == T @ uplo)

[Package Matrix version 0.999375-30 Index]