expand {Matrix}	R Documentation

Expand a Decomposition into Factors

Description

Expands decompositions stored in compact form into factors.

Usage

expand(x, ...)

Arguments

x 	 a matrix decomposition.

... 	 further arguments passed to or from other methods.

Details

This is a generic function with special methods for different types of decompositions, see showMethods(expand) to list them all.

Value

The expanded decomposition, typically a list of matrix factors.

Note

Factors for decompositions such as lu and qr can be stored in a compact form. The function expand allows all factors to be fully expanded.

See Also

The LU decomposition lu, for which there is an expand method; facmul.

Examples

(x <- Matrix(round(rnorm(9),2), 3, 3))
(ex <- expand(lux <- lu(x)))

[Package Matrix version 0.999375-30 Index]