image-methods {Matrix}	R Documentation

Methods for image() in Package 'Matrix'

Description

Methods for function image in package Matrix. An image of a matrix simply color codes all matrix entries and draws the n x m matrix using an n x m grid of (colored) rectangles.

Usage

## S4 method for signature 'dgTMatrix':
image(x,
      xlim = .5 + c(0, di[2]),
      ylim = .5 + c(di[1], 0), aspect = "iso",
      sub = sprintf("Dimensions: %d x %d", di[1], di[2]),
      xlab = "Column", ylab = "Row",
      useAbs = NULL, colorkey = !useAbs,
      col.regions = NULL,
      lwd = NULL, ...)

Arguments

x 	 a Matrix object, i.e., fulfilling is(x, "Matrix").

xlim, ylim 	 x- and y-axis limits; may be used to “zoom into” matrix.

aspect 	 aspect ratio specified as number (y/x) or string; see levelplot.

sub, xlab, ylab 	 axis annotation with sensible defaults; see plot.default.

useAbs 	 logical indicating if abs(x) should be shown; if TRUE, the former (implicit) default, the default col.regions will be grey colors (and no colorkey drawn). The default is FALSE unless the matrix has no negative entries.

colorkey 	 logical indicating if a color key aka ‘legend’ should be produced. Default is to draw one, unless useAbs is true.

col.regions 	 vector of gradually varying colors; see levelplot.

lwd 	 non-negative number or NULL (default), specifying the line-width of the rectangles of each non-zero matrix entry (drawn by grid.rect). The default depends on the matrix dimension and the device size.

... 	 further arguments passed to methods and levelplot.

Methods

All methods currently end up calling the method for the dgTMatrix class. Use showMethods(image) to list them all.

Examples

showMethods(image)
## If you want to see all the methods' implementations:
showMethods(image, incl=TRUE, inherit=FALSE)

data(CAex)
image(CAex, main = "image(CAex)")
image(CAex, useAbs=TRUE, main = "image(CAex, useAbs=TRUE)")

cCA <- Cholesky(crossprod(CAex), Imult = .01)
image(cCA, main="Cholesky(crossprod(CAex), Imult = .01)")
image(cCA, useAbs=TRUE)

data(USCounties)
image(USCounties)# huge
image(sign(USCounties))## just the pattern
    # how the result looks, may depend heavily on
    # the device, screen resolution, antialiasing etc
    # e.g. x11(type="Xlib") may show very differently than cairo-based

## Drawing borders around each rectangle;
    # again, viewing depends very much on the device:
image(USCounties[1:400,1:200], lwd=.1)
image(USCounties[1:200,1:200], lwd=.5)
image(USCounties[1:200,1:200], lwd=.01)

[Package Matrix version 0.999375-30 Index]