isSymmetric-methods {Matrix}	R Documentation

Methods for Function isSymmetric in Package 'Matrix'

Description

isSymmetric(M) returns a logical indicating M is a symmetric matrix. The methods are defined for virtual Matrix classes such that it works for all objects inheriting from class Matrix.

Methods

object = "symmetricMatrix"

...

object = "triangularMatrix"

...

object = "denseMatrix"

...

object = "diagonalMatrix"

...

object = "sparseMatrix"

...

[Package Matrix version 0.999375-30 Index]