ndenseMatrix-class {Matrix}	R Documentation

Virtual Class "ndenseMatrix" of Dense Logical Matrices

Description

ndenseMatrix is the virtual class of all dense logical (S4) matrices. It extends both denseMatrix and lMatrix directly.

Slots

x:

logical vector containing the entries of the matrix.

Dim, Dimnames:

see Matrix.

Extends

Class "lMatrix", directly. Class "denseMatrix", directly. Class "Matrix", by class "lMatrix". Class "Matrix", by class "denseMatrix".

Methods

%*%

signature(x = "nsparseMatrix", y = "ndenseMatrix"): ...

%*%

signature(x = "ndenseMatrix", y = "nsparseMatrix"): ...

coerce

signature(from = "matrix", to = "ndenseMatrix"): ...

coerce

signature(from = "ndenseMatrix", to = "matrix"): ...

crossprod

signature(x = "nsparseMatrix", y = "ndenseMatrix"): ...

crossprod

signature(x = "ndenseMatrix", y = "nsparseMatrix"): ...

as.vector

signature(x = "ndenseMatrix", mode = "missing"): ...

diag

signature(x = "ndenseMatrix"): extracts the diagonal as for all matrices, see the generic diag().

See Also

Class ngeMatrix and the other subclasses.

Examples

showClass("ndenseMatrix")

as(diag(3) > 0, "ndenseMatrix")# -> "nge"

[Package Matrix version 0.999375-30 Index]