number-class {Matrix}	R Documentation

Class "number" of Possibly Complex Numbers

Description

The class "number" is a virtual class, currently used for vectors of eigen values which can be "numeric" or "complex".

It is a simple class union (setClassUnion) of "numeric" and "complex".

Objects from the Class

Since it is a virtual Class, no objects may be created from it.

Examples

showClass("number")
stopifnot( is(1i, "number"), is(pi, "number"), is(1:3, "number") )

[Package Matrix version 1.3-3 Index]