packedMatrix-class {Matrix}	R Documentation

Virtual Class "packedMatrix" of Packed Dense Matrices

Description

Class "packedMatrix" is the virtual class of dense symmetric or triangular matrices in "packed" format, storing only the choose(n+1,2) == n*(n+1)/2 elements of the upper or lower triangle of an n-by-n matrix. It is used to define common methods for efficient subsetting, transposing, etc. of its proper subclasses: currently "[dln]spMatrix" (packed symmetric), "[dln]tpMatrix" (packed triangular), and subclasses of these, such as "dppMatrix", "pCholesky", and "pBunchKaufman".

Slots

uplo:

"character"; either "U", for upper triangular, and "L", for lower.

Dim, Dimnames:

as all Matrix objects.

Extends

Class "denseMatrix", directly. Class "Matrix", by class "denseMatrix", distance 2. Class "mMatrix", by class "Matrix", distance 3. Class "replValueSp", by class "Matrix", distance 3.

Methods

pack

signature(x = "packedMatrix"): ...

unpack

signature(x = "packedMatrix"): ...

isSymmetric

signature(object = "packedMatrix"): ...

isTriangular

signature(object = "packedMatrix"): ...

isDiagonal

signature(object = "packedMatrix"): ...

t

signature(x = "packedMatrix"): ...

diag

signature(x = "packedMatrix"): ...

diag<-

signature(x = "packedMatrix"): ...

[

signature(x = "packedMatrix", i = "missing", j = "missing", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "NULL", j = "missing", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "index", j = "missing", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "matrix", j = "missing", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "lMatrix", j = "missing", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "missing", j = "NULL", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "NULL", j = "NULL", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "index", j = "NULL", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "matrix", j = "NULL", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "lMatrix", j = "NULL", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "missing", j = "index", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "NULL", j = "index", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "index", j = "index", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "matrix", j = "index", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "lMatrix", j = "index", drop = "missing"): ...

[

signature(x = "packedMatrix", i = "missing", j = "missing", drop = "logical"): ...

[

signature(x = "packedMatrix", i = "NULL", j = "missing", drop = "logical"): ...

[

signature(x = "packedMatrix", i = "index", j = "missing", drop = "logical"): ...

[

signature(x = "packedMatrix", i = "matrix", j = "missing", drop = "logical"): ...

[

signature(x = "packedMatrix", i = "lMatrix", j = "missing", drop = "logical"): ...

[

signature(x = "packedMatrix", i = "missing", j = "NULL", drop = "logical"): ...

[

signature(x = "packedMatrix", i = "NULL", j = "NULL", drop = "logical"): ...

[

signature(x = "packedMatrix", i = "index", j = "NULL", drop = "logical"): ...

[

signature(x = "packedMatrix", i = "matrix", j = "NULL", drop = "logical"): ...

[

signature(x = "packedMatrix", i = "lMatrix", j = "NULL", drop = "logical"): ...

[

signature(x = "packedMatrix", i = "missing", j = "index", drop = "logical"): ...

[

signature(x = "packedMatrix", i = "NULL", j = "index", drop = "logical"): ...

[

signature(x = "packedMatrix", i = "index", j = "index", drop = "logical"): ...

[

signature(x = "packedMatrix", i = "matrix", j = "index", drop = "logical"): ...

[

signature(x = "packedMatrix", i = "lMatrix", j = "index", drop = "logical"): ...

Author(s)

Mikael Jagan

See Also

pack and unpack; its virtual "complement" "unpackedMatrix"; its proper subclasses "dspMatrix", "ltpMatrix", etc.

Examples

showClass("packedMatrix")
showMethods(classes = "packedMatrix")

[Package Matrix version 1.5-3 Index]