qr-methods {Matrix}	R Documentation

QR Decomposition – S4 Methods and Generic

Description

The "Matrix" package provides methods for the QR decomposition of special classes of matrices. There is a generic function which uses qr as default, but methods defined in this package can take extra arguments. In particular there is an option for determining a fill-reducing permutation of the columns of a sparse, rectangular matrix.

Methods

x = "dgCMatrix"

QR decomposition of a general sparse double-precision matrix with nrow(x) >= ncol(x). Returns an object of class "sparseQR".

x = "sparseMatrix"

works via "dgCMatrix".

See Also

qr; documentation for classes sparseQR and dgCMatrix.

[Package Matrix version 0.999375-30 Index]