replValue-class {Matrix}	R Documentation

Virtual Class "replValue" - Simple Class for Subassignment Values

Description

The class "replValue" is a virtual class used for values in signatures for sub-assignment of Matrix matrices.

In fact, it is a simple class union (setClassUnion) of "numeric" and "logical" (and maybe "complex" in the future).

Objects from the Class

Since it is a virtual Class, no objects may be created from it.

See Also

Subassign-methods, also for examples.

Examples

showClass("replValue")

[Package Matrix version 1.6-2 Index]