unpack {Matrix}	R Documentation

Full Storage Representation of Packed Matrices

Description

Expands decompositions stored in compact form into matrix factors.

Usage

unpack(x, ...)

Arguments

x 	 a matrix stored in packed form, e.g., of class "d?pMatrix" where "?" is "t" for triangular or "s" for symmetric.

... 	 further arguments passed to or from other methods.

Details

This is a generic function with special methods for different types of packed matrices. Use showMethods("unpack") to list the methods for unpack().

Value

A Matrix object containing the full-storage representation of x.

Examples

showMethods("unpack")
(cp4 <- chol(Hilbert(4))) # is triangular
tp4 <- as(cp4,"dtpMatrix")# [t]riangular [p]acked
str(tp4)
(unpack(tp4))

## Not run: 
## once we have Diagonal() :
unpack(Diagonal( 1:3))
## End(Not run)

[Package Matrix version 0.999375-30 Index]